More experimentation with modifiers
Now moved back modifiers to the base layer thumbs, and changed tab
places, hoping to get the modifiers more accessible even if in the inner
thumbs which are annoying to hit.

As a con, the delete button has to be on the num layer now. Will see how
that feels.